+++ CHANGELOG.md
@@ -8,6 +8,10 @@
   - Fix the problem of duplicate detection of the same SV candidate
 - Add index to ensure uniqueness of evidence bam filenames (MANTA-1431)
   - It solves the potential problem of name conflicts for evidence bams if the input bam files have the same name while located in different directories.
+- Change filters for easy interpretation of multi-sample germline variant vcf
+  - Add record-level filter 'SampleFT' when no sample passes all sample level filters
+  - Add sample-level filter 'HomRef' for homogyzous reference calls
+  - No more sample-level filter will be applied at the record level even if it applies to all samples
 
 ## v1.4.0 - 2018-04-25